Divide functional test by class
